requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
EbookLib==0.18
fire==0.6.0
idna==3.7
iniconfig==2.0.0
jieba==0.42.1
lxml==5.2.2
maturin==1.6.0
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
pytest==8.2.2
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
termcolor==2.4.0
urllib3==2.2.1
